Move base exceptions to c-c-base
