+++ platforms/core-configuration/configuration-cache-base/src/main/kotlin/org/gradle/internal/cc/base/exceptions/Exceptions.kt
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.cc.base.exceptions
+
+import org.gradle.internal.exceptions.Contextual
+import org.gradle.internal.exceptions.DefaultMultiCauseException
+
+/**
+ * Marker interface for exception handling.
+ */
+interface ConfigurationCacheThrowable
+
+/**
+ * State might be corrupted and should be discarded.
+ */
+@Contextual
+class ConfigurationCacheError(
+    error: String,
+    cause: Throwable? = null
+) : ConfigurationCacheThrowable, Exception(error, cause)
+
+@Contextual
+open class ConfigurationCacheException(
+    message: () -> String,
+    causes: Iterable<Throwable>
+) : DefaultMultiCauseException(message, causes), ConfigurationCacheThrowable
+
